Fix some clippy warnings (#34)
* fix clippy warning: "result_large_err"

The servererror variant of neo4jerror is huge. on my computer it is 184
bytes. clippy warns us that it's not recommended to have an err variant
that big. the solution is to box the servererror, which requires a memory
allocation, but reduces the size of the result.
this way, we won't need to reserve that much stack space for every result.
errors are less common to occur, so the benefit of reducing the size of
all results should compensate for the memory allocation required in case
of a servererror.

see the details of this lint on the link below:
https://rust-lang.github.io/rust-clippy/master/index.html#result_large_err>

* Fix clippy warning: "unneeded_struct_pattern"

The RecordBuffer::Transaction variant is empty. Clippy suggests
removing the struct pattern `{..}`.

* Avoid unnecessary memory allocation on driver::io::bolt::handshake

Instead of using the format! macro, which creates a string, feed the
format directly to the trace! and debug! macros.
This achieves the same result, without creating an intermediate String.

* Fix more clippy lints

 * inline formatting arguments
 * avoid `mem::swap` into temporary
 * avoid `.map(|x| x.cloned())`

---------

Co-authored-by: Robsdedude <dev@rouvenbauer.de>