Improve handling of cluster network partitioning (#43)
In case of a network partitioning in the cluster, the minority will not be able
to elect a leader and therefore not be able to server write queries (besides,
its state will fall behind leading to stale reads).

This PR changes the driver to keep looking for routing tales if it receives one
without a writer as this might be (but not necessarily is) the symptom of a
network partitioning.